Quik/backend/requirementes.txt
Django==5.2.7
djangorestframework==3.16.1
asgiref==3.10.0
sqlparse==0.5.3
pytz==2022.1
mysql-connector-python==9.1.0